+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/__init__.py
@@ -55,9 +55,8 @@ from ._models import (
     RecognizeCustomEntitiesAction,
     RecognizeCustomEntitiesResult,
     SingleCategoryClassifyAction,
-    SingleCategoryClassifyResult,
     MultiCategoryClassifyAction,
-    MultiCategoryClassifyResult,
+    ClassifyDocumentResult,
     ClassificationCategory,
     AnalyzeHealthcareEntitiesAction,
 )
@@ -117,9 +116,8 @@ __all__ = [
     "RecognizeCustomEntitiesAction",
     "RecognizeCustomEntitiesResult",
     "SingleCategoryClassifyAction",
-    "SingleCategoryClassifyResult",
     "MultiCategoryClassifyAction",
-    "MultiCategoryClassifyResult",
+    "ClassifyDocumentResult",
     "ClassificationCategory",
     "AnalyzeHealthcareEntitiesAction",
 ]