+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/_text_analytics_client.py
@@ -63,9 +63,8 @@ from ._models import (
     RecognizeCustomEntitiesAction,
     RecognizeCustomEntitiesResult,
     SingleCategoryClassifyAction,
-    SingleCategoryClassifyResult,
     MultiCategoryClassifyAction,
-    MultiCategoryClassifyResult,
+    ClassifyDocumentResult,
     AnalyzeHealthcareEntitiesAction,
     _AnalyzeActionsType,
 )
@@ -85,8 +84,7 @@ AnalyzeActionsResponse = AnalyzeActionsLROPoller[
                 AnalyzeSentimentResult,
                 ExtractSummaryResult,
                 RecognizeCustomEntitiesResult,
-                SingleCategoryClassifyResult,
-                MultiCategoryClassifyResult,
+                ClassifyDocumentResult,
                 AnalyzeHealthcareEntitiesResult,
                 DocumentError,
             ]
@@ -1091,8 +1089,7 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
                     AnalyzeSentimentResult,
                     ExtractSummaryResult,
                     RecognizeCustomEntitiesResult,
-                    SingleCategoryClassifyResult,
-                    MultiCategoryClassifyResult,
+                    ClassifyDocumentResult,
                     AnalyzeHealthcareEntitiesResult,
                     DocumentError,
                 ]
@@ -1151,8 +1148,7 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
             ~azure.ai.textanalytics.AnalyzeActionsLROPoller[~azure.core.paging.ItemPaged[
             list[RecognizeEntitiesResult or RecognizeLinkedEntitiesResult or RecognizePiiEntitiesResult or
             ExtractKeyPhrasesResult or AnalyzeSentimentResult or ExtractSummaryAction or RecognizeCustomEntitiesResult
-            or SingleCategoryClassifyResult or MultiCategoryClassifyResult or AnalyzeHealthcareEntitiesResult or
-            DocumentError]]]
+            or ClassifyDocumentResult or AnalyzeHealthcareEntitiesResult or DocumentError]]]
         :raises ~azure.core.exceptions.HttpResponseError or TypeError or ValueError:
 
         .. versionadded:: v3.1
@@ -1160,8 +1156,8 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
         .. versionadded:: 2022-04-01-preview
             The *ExtractSummaryAction*, *RecognizeCustomEntitiesAction*, *SingleCategoryClassifyAction*,
             *MultiCategoryClassifyAction*, and *AnalyzeHealthcareEntitiesAction* input options and the
-            corresponding *ExtractSummaryResult*, *RecognizeCustomEntitiesResult*, *SingleCategoryClassifyResult*,
-            *MultiCategoryClassifyResult*, and *AnalyzeHealthcareEntitiesResult* result objects
+            corresponding *ExtractSummaryResult*, *RecognizeCustomEntitiesResult*, *ClassifyDocumentResult*,
+            and *AnalyzeHealthcareEntitiesResult* result objects
 
         .. admonition:: Example: