+++ sdk/textanalytics/azure-ai-textanalytics/tests/test_analyze_async.py
@@ -42,8 +42,7 @@ from azure.ai.textanalytics import (
     SingleCategoryClassifyAction,
     MultiCategoryClassifyAction,
     RecognizeCustomEntitiesAction,
-    SingleCategoryClassifyResult,
-    MultiCategoryClassifyResult,
+    ClassifyDocumentResult,
     RecognizeCustomEntitiesResult,
     AnalyzeHealthcareEntitiesAction
 )
@@ -1175,8 +1174,9 @@ class TestAnalyzeAsync(TextAnalyticsTest):
                 assert not result.is_error
                 assert not result.warnings
                 assert result.statistics
-                assert result.classification.category
-                assert result.classification.confidence_score
+                for classification in result.classifications:
+                    assert classification.category
+                    assert classification.confidence_score
 
     @pytest.mark.skipif(not is_public_cloud(), reason='Usgov and China Cloud are not supported')
     @TextAnalyticsCustomPreparer()
@@ -1317,8 +1317,8 @@ class TestAnalyzeAsync(TextAnalyticsTest):
                 document_results.append(doc)
 
         assert len(document_results) == 2
-        assert isinstance(document_results[0][0], SingleCategoryClassifyResult)
-        assert isinstance(document_results[0][1], MultiCategoryClassifyResult)
+        assert isinstance(document_results[0][0], ClassifyDocumentResult)
+        assert isinstance(document_results[0][1], ClassifyDocumentResult)
         assert isinstance(document_results[0][2], RecognizeCustomEntitiesResult)
         assert document_results[1][0].is_error
         assert document_results[1][1].is_error